Implement sdp-shutdown using an external service to do the shutdown
The external service must be registered in consul (with the name
"poweroff") and support a POST to /poweroff to trigger the shutdown. The
master controller queries consul for all machines providing the service
and pokes them. This replaces the old method of launching a process on
every slave, which is much less convenient now that the master
controller doesn't talk directly with Mesos. Possibly it could have been
replaced with code to interact with Singularity, but this approach has
some advantages:
- It can shut down machines that aren't Mesos slaves (or even have
  Docker).
- It can shut down machines that are 100% loaded with Mesos tasks.
- It should be faster, since there is no need to wait for images to be
  pulled to the slaves while the data centre is on fire.

It does require the machines to be part of a Consul cluster and to run
an internally-written service.