Update the rotator to work inside master controller
There were a few changes to the websocket interface since it can use the
same information dictionary that is used to drive the index of GUIs.

A side effect is that it now requires ES6 because it's the first version
that actually has a sane interface for iterating over arrays.