Use random port assignments
The goal is to avoid a port being reassigned to another service too
quickly, which can be a problem if there is a user still trying to
connect to the old port (e.g. dashboard AJAX requests).

The previous scheme was to allocate ports sequentially, wrapping back to
the start. However, that requires global information to be useful, which
is no longer available in the product controller's scheduler.

Closes SPR1-163.