Fix display of batch task details
It was previously showing details for the realtime task in the same row.

Also now uses row IDs to ensure that it's showing the information for
the right task, even if the set of tasks changes between populating the
table and getting details for a row of the table.

Fixes SPR1-170.