Report start time of products in ?product-list
This became a bit complicated because it was the first time the zk_state
schema version has been bumped. The schema was converted to use the
same multi-versioning system the product_config schema uses. The system
then needed to be modified to cope with version being an integer rather
than a string. Now instead of the full schema being in the template body
(which required some valid version to be supplied just to load the
template), the full schema is in a macro parametrised by the version.

Fixes SPR1-167.