Plumb in device-status
The product controller now sets the device-status to FAIL if the
subarray product goes into ERROR. Note that it never comes out of FAIL,
whereas the product state will change to DECONFIGURING on deconfigure.

The master controller's device-status sensor reflects the worst
device-status of any product. This could still be improved by checking
for failures of internal components (e.g. if reconciliation fails it
should set a degraded status). See SPR1-152.

The master controller now uses SensorWatcher directly instead of
SensorProxyClient. That wasn't strictly a necessary change (it's mostly
the product of me refactoring how I wanted to implement this), but I
think it makes it clearer that there are two separate sensor watchers
involved.

This depends on a bugfix in aiokatcp 0.6.1 for reliable operation
(otherwise a stale value for the product's device-status may be used
by the master due to the sensor watchers running in arbitrary order).