Remove nvidia-docker 1 support: part 1
This removes the master controller code that checked whether
nvidia-container-runtime was present, and now just assumes it is
present. It also makes `uuid` a required attribute of a GPU, and
`driver_version` optional (it's no longer needed for anything, but still
provided).

Step 2 will involve removing support from agent_mkconfig, which can only
be done once this PR is used in all running master controllers. Step 3
is to tighten the schema to remove the fields that were still being
permitted for compatibility, and can only be done once all the agent
machines have been updated.