Take max_channels into account in computing writer chunking
This will give larger chunks for the spectral imager data stream,
hopefully improving performance (closes SPR1-227). The max memory usage
for the writers is increased to 32GB. In practice I think that's only
going to affect large 32K-channel arrays when writing visibilities for
spectral imaging, because in other cases either the time length or the
chunk size limit kicks in first.