Improve overriding image path for ingest
The previous scheme first ran the normal lookup with
katsdpingest_titanx, then replaced it by looking up the image for the
actual GPU (if different). That caused two problems:

1. If katsdpingest_titanx didn't exist, it would fail, even though the
image wouldn't be needed.
2. The .version sensor indicating the image in use would be wrong, since
it would point to the looked-up katsdpingest_titanx image.

Instead, we allow the low-level image lookup to be overridden. in
`resolve`, and have the sub-class `resolve` pass an override along the
chain of `super` calls.