Make 'Lost connection' display on dashboard work again
It was being thwarted by the proxying: it redirected to the master
controller's web server, and even though that returned a 405, dash would
treat this as completion and reset the loading state.

Hack around that by delaying the response by 2s, so that dash always
believes it has stale data.

Closes SPR1-257.