Limit imaging to observations with enough time on target
This required some restructuring of the "target cache". Previously it
cached full target list, but spectral and continuum no longer image the
same set of targets. It's been replaced by a `TargetMapper` that is
responsible just for assigning a name to a target in an idempotent way.
This actually simplifies the interfaces since `_get_targets` (formerly
known as `_unique_targets`) no longer has any responsibility for naming.