Enable haproxy statistics page, plus health checks
It adds a /stats to the server, from which we can see the status of the
various backend servers. I've also enabled TCP-level health checks so
that we can see whether haproxy is able to connect to the servers. I
briefly tried http health checks, but most of the backend servers return
a 404 if asked for / so that failed (we could of course try asking for
the URL where they actually display the dashboard, but in some cases I
think that could be expensive).