Explicitly set haproxy ulimit-n (again)
The auto-computed limit depends on the number of backends. If it was
started with no arrays then it would set a low limit, and warn because
it was unable to raise it when a new backend was added. That's only
serious if we actually get close to maxconn, but for safety I'm now
setting a ulimit-n that's much higher than the auto-computed value and
should handle hundreds of backends.