Build haproxy from source
The PPA I was using released a new version and **deleted** the version
we've pinned. So rather than have the carpet removed from under us
every release, I'm building haproxy from a source download. This might
even make the image slightly smaller since only the binary is copied to
the final image, not the docs.