Give spectral imager more RAM
The calculation is now based on the number of visibilities to be loaded
(as they are held in memory). The current value is rather conservative,
and we can fine-tune it later. It's possible that for very long
observations it will lead to a loss of parallelism if it needs more than
32GB per task (since each imager has 4 GPUs and 128 GB of RAM). That
will occur at about 6-7 hours of full-size 32K, 8s data.