Support separate URLs for read and write in s3_config
For compatibility, each archive can be specified as before, or url's can
be specified separately in the `read` and `write` sections and not in
the top level (making for a somewhat complex JSON schema). There is a
schema fix, where previously the access key and secret key weren't
marked required even though they are.

Internally, the old format is rewritten into the new format by the
product controller when it starts up.

This is slightly brittle in terms of backwards/forwards compatibility:
once we update the config file, we'll only be able to use the new
version of the product controller. It'll also need a new
katsdpdatawriter to actually have the desired effect, although an old
one will still work but use the read endpoint for writing.