More fixups for split S3
I rewrote the schema for s3_config, and updated to a newer version of
jsonschema that supports if-then. The "archive" section no longer needs
a "write" section, but does require a read URL (either in a "read"
section - which need not have keys - or at the top level).

The test and sandbox s3_config's now use this more minimal archive
section, mostly to catch errors where the other information is expected.

When pushing urls down in `_normalise_s3_config`, read/write sections
will be created if necessary to push URLs into.