Escape computed command-line arguments
Arguments in the `command` attribute of tasks are passed to `str.format`
to allow extra properties (mostly from the resolver) to be filled in at
resolution time. When a parameter is a variable rather than a literal,
there is a risk that it could contain a `{` which would break the
formatting. Variable string parameters are now passed to an escaping
function.

Closes SPR1-362.