Make unit tests reliable again
PR #466 changed the handling of failures of dependencies in a way that
required some mocking in the unit tests, but the way it was handled
wasn't reliable (basically, if the failure didn't propagate and set the
state to KILLING before the mocking code inspected the state, the
mocking code wouldn't set the state to DEAD). The approach to mocking
has been changed so that the KILLING -> DEAD transition is handled in a
wrapper around `dependency_abort`, which seems to be more reliable.