Add depends_finished_critical edge property
When explicitly set to false, a depends_finished edge will force
ordering, but the downstream task will run even when the upstream task
fails. This is to support a future spectral imager report, which must
run after the result of spectral imaging but should run even if
individual spectral tasks fail.

Relates to SPR1-108.