Bumble on with postprocessing
If a capture-done fails, or one of the postprocessing transitions fails
(which I think is just the capture-done for the flag writer), carry on
anyway with postprocessing.

Fixes SPR1-358. It possibly introduces much worse issues in the corner
cases, particularly if cal jams up while writing the report and hence
never finishes with the capture block.