Fix an obscure test error with Python 3.8
There seems to be some subtle changes in how side_effect is handled for
autospeced mocks of async methods: the mock knows it has to be
asynchronous, so asynchronously returns the result of the side_effect.
If the side_effect is a coroutine function that is handled sensibly, but
if it is a regular function returning a future then the future is
returned asynchronously (making things annoyingly double-asynchronous).