Stop publishing katcp sensors for batch tasks
There were .version, .state and .mesos-state sensors being published,
but this has caused problems for CAM which doesn't cope well with a
large number of sensors combined with a high frequency of
interface-changed notifications.

It also adds a missing interface-changed notification when a task with
these sensors starts up.

This should mitigate MKAIV-1593.