Double product controller memory again
To fix SPR1-614. It's probably not necessary, but it's still not a huge
amount of memory and when the product controller OOMs it is not pretty.

It would have been nicer to scale this according to the number of
channels, but we don't have that information at the time the product
controller is spun up.