Optimise getting sensors from katportal
katportal on MeerKAT is very slow (takes 1.5s+ to get each sensor). I'm
guessng it lacks a dictionary lookup and iterates over every single
sensor. Speed it up by munging all the sensors to query into a regex and
querying with that, which at least only makes one pass over the sensor
list on the backend.

This should fix SPR1-621.