Better error when launch times out in queue
If a launch with a lower resources_timeout is behind one with a higher
resources_timeout, it can time out even without reaching the front of
the queue. In this case it was giving the not-entirely-correct message
"No resource offers received", even if the launch doesn't have any
physical tasks and hence doesn't need resources.