Make Mesos responsible for timing out batch tasks
This simplifies the code slightly, and ensures that tasks will be killed
off even if there is a network partition or the product controller gets
wedged.

Some extra information is added to TaskError messages to report the
reason (which for these timeouts is REASON_MAX_COMPLETION_TIME_REACHED)
so that it will be clear why the task failed.

Closes SPR1-488.