Fix product controller after interface mode cleanup
7b9c0af removed the code that passed interface_mode to DeviceServer, but
it's a required argument. Add it back.

Also remove sandbox/sdp_image_tag, because it's local configuration and
if it's in version control, it would be very easy for someone to test on
a branch, accidentally check in the file, and then someone else will
pull down the change and end up working with a different tag to the one
they thought they were. Add it to .gitignore instead.