Put image overrides through the lookup machinery
Previously, if an image was overridden, say from foo to
sdp-docker-registry.kat.ac.za:5000/foo:my-branch, that would be passed
directly into the Mesos taskinfo, and without a force_pull_image. If the
Mesos agent had an older version of that image it would be used rather
than pulling a new version. In practice that meant one had to manually
look up the hash and use that in the override.

Instead, the overridden value is put through the ImageLookup machinery,
meaning that (unless `--no-pull` is given), the master controller will
query the registry for the latest hash the first time the image is used
for a subarray.

A side benefit is that it's not necessary to give a full path: one can
override foo -> foo:my-branch and it will use the default private
registry.