Support token authentication with a registry
This will allow katsdpcontroller to work with harbor as the registry (or
any registry implementing
https://docs.docker.com/registry/spec/auth/token/). When an HTTP 401
response is receive with a suitable WWW-Authenticate, send another
request to that token server to get a token and then retry the original
request with that token.

This also contains a bug fix: if an explicit registry is provided for a
request, look up the appropriate username and password for that
registry, rather than using the credentials for the default registry.

The unit test has also been significantly beefed up, to test a number of
failure cases, and to actually check that credentials are being looked
up and passed correctly (which also required fixing the test to store
them in the right format).

Closes NGC-389.