Decrease number of substreams for gpucbf correlator
With 8KiB packets, xbgpu can handle much more data than a SKARAB
X-engine, so we need fewer of them. However, we split the threads across
two cores as one core might not be able to quite take it. CPU usage is
set to 1.5 cores since the cores are not fully utilised.

See NGC-354.