Tweaks to linkage with consul
The consul `Name` is changed to be the task type: this means that consul
service discovery can be used to find all F-engine tasks for example.
The specific instance e.g. `f.antenna_channelised_voltage.1` is put into
a `task_name` metadata field.

The sandbox Prometheus config is also adjusted: the task type is now
mapped to `job` (a standard Prometheus label).