Adjust memory usage for xbgpu
Set `--heaps-per-fengine-per-chunk` to aim for a 16MB chunk size, to
ensure chunks are not so small that per-chunk overheads become
significant, while also limiting the memory needed for the visibility
accumulators (which scale with the number of batches per chunk).

At the same time, try to more accurately model the memory usage of
xbgpu, both on the host and GPU, and update the GPU compute allocation.

Relates to NGC-526.