Implement gain-all command
This is larger than might be expected because there is also some
refactoring to simplify finding the katcp servers corresponding to a
stream. Each SDPLogicalTask may now have a set of associated stream
names, and a new SDPSubarrayProduct.find_nodes method can use this to
find the nodes for a stream. There is also a `_multi_request` helper for
issuing multiple katcp requests in parallel.

This still needs unit tests. It also won't actually work because
katgpucbf does not yet implement the request.

See NGC-498.