Add support for "subsystems"
This will allow some machines to be reserved for SDP and others for CBF,
so that each can evaluate their respective hardware without tasks
wandering between them.

I would have preferred to do this with Mesos roles since that would
allow more flexibility (e.g. could partition a single machine), but
trying to launch a single subarray across multiple roles seemed fraught
with complication. I think Mesos would offer shared resources to one or
the other role, and the scheduler would need to be smart enough to
decline it where necessary to allow them to be offered again with the
other role.