Eliminate nosetests completely
Rather than trying to write something similar to assert_count_equal (or
using unittest.TestCase.assertCountEqual) I just tweaked the tests to
use equality comparison and ensured the expected value was in the right
order. This makes the test slightly more fragile, but that area of code
hasn't changed for years so I don't see it being an issue in practice.