Make things work with async_timeout 4.0
The tests broke because an explicit loop argument was being used. There
was also one place that it was being used with the synchronous context
manager, which is deprecated.

This will be superseded by #599, but that's a large PR and may take some
time to land, and in the meantime the master branch is not building.