Implement capture-list for CBF
To implement it, each PhysicalMulticast object has an enum to indicate
whether it is up, down, or unknown. Unknown streams won't appear in
capture-list.

It only lists gpucbf streams, because reporting the up/down status
requires tracking whether streams are up or down, and I didn't feel like
trying to do that for all the assorted SDP streams. It's not part of the
SDP ICD, so there is no need to do that.

Closes NGC-716.