Split develop mode into finer options
Develop mode currently changes the product_configuration in three ways:
1.) It allows any gpu to be used for the ingest process
2.) It switches off ibverbs and no longer requests an infiniband
connection for the cbf network
3.) It reduces the resource allocation for some processes to allow a
developer to run the system on their laptop. It also modifies the number
of servers created for ingest and calibration.

This PR allows a user to select only some of these
features/modifications when running develop mode.