Fix more pytest deprecation warnings
Where a `setup` method was just creating one fixture I turned it into a
fixture; in more complex cases I just renamed it to `setup_method` to
keep pytest happy.

With suitable updates to dependencies, this reduces the warnings to just
one from the `past` package (part of `future`), which is no longer
maintained. Hopefully that will cease to be an issue once katpoint 1.0
is released.