Add hardcoded ddc-mix-freq sensor
For the time being, this just reads zero, to fulfil the obligations that
we have as per the ICD (NGC-831).

I think ideally, the value should be read from the stream configuration,
and I did initially try poking at that a bit, but I don't understand
enough about how this works to be able to be sure I haven't borken
something.

So perhaps the proper logic should be introduced as part of NGC-564 or
NGC-565.