Add .source and .revision sensors for product controller
This required updating the Zookeeper persistence schema to store all the
information about an image. If an existing product controller is running
using the old schema, its image information won't be loaded, but I think
the only implication of that is that the sensors won't be populated; it
can still be controlled.