Use fullstops for hierarchical separation in sensor names
This change affects the product controller. Previously, hyphens were
used basically everywhere.

The CBF ICD with CAM requires that the KATCP guidelines (i.e. using a
fullstop to separate hierarchical levels), which is what prompted this
change.

Note: there are some not-exclusively-CBF sensors that are affected here.
I'm not sure whether or how much this will affect SDP.

Addresses NGC-854.