SPR1-2134
This PR implements suggestion 2 in the SPR1-2134 ticket.
Suggestion 1 will be implemented in subsequent PR.

Develop mode currently changes the product_configuration in three ways:
1.) It allows any gpu to be used for ingest
2.) It switches off ibverbs and no longer requests an infiniband
connection for the cbf network
3.) It reduces the resource allocation for some processes to allow a
developer to run the system on their laptop. It also modifies the number
of servers created for ingest and calibration.

This PR allows a user to select only some of these modifications when running develop mode,
instead of the whole set.
This can be useful for testing in the lab environment where for example
you would like to use fewer resources than in the production environment
but still enable ibverb support for performance testing.