Control autotune matching behavior in ingest tasks
This commit sets the KATSDPSIGPROC_TUNE_MATCH env variable inside
ingest containers before launching. This controls the behaviour
of the autotuning machinery when an exact match for the GPU device
is not found. Setting the variable to "nearest" causes the nearest
match to be used, if any exists, while setting it to "exact" forces
autotuning for all but exact matches.