SPR1-2376 Don't dump RDB files if not archiving
While the meta_writer's `write_meta` KATCP request is quite flexible
in terms of stream input, it is only ever called with the default
setting, which looks for archived sdp.vis streams. The resulting stream
names are appended to the CBID to form corresponding RDB filenames.

When testing in the lab, we frequently disable archiving streams to
save disk space. The side effect is that meta_writer crashes with "No
stream specified, and cannot determine available streams from telstate."

Avoid this crash by disabling RDB dumps if there are no archived
streams. This is necessary but not sufficient, but should at least
get rid of the typical lab crashes.