SPR1-2376 Let dsim set sub_band telstate entry
In the absence of CAM we need to set certain telstate keys as part of
product configuration. One is `sub_band`, needed by cal to start up.
The CBF simulator stream ("sim.cbf") already initialises `sub_band`,
so do the same for the lab test setup of the GPU correlator.

Dsim is a good place to put this because
  - it knows the band,
  - it already pre-populates `mxxx_observer` in the absence of CAM, and
  - it is typically not needed if CAM is around.