Support using skopeo for copying images to sandbox
The standard process involves first copying the image to the local
Docker registry (with `docker pull`) then tagging it and pushing it
(with `docker push`). That leads to a lot of unwanted tagged images in
one's local Docker engine.

Skopeo is a tool that directly copies from one registry to another,
without involving the Docker daemon. For now it's optional.