Don't keep product sensors around in master controller
The previous commit changed sensor_proxy.py so that sensors persist but
are marked unreachable if a connection is closed. That was desired for
task sensors reflected in a product, but when I product goes away, the
sensors should vanish too.

Make that work by extending SensorWatcher with a field to indicate what
to do with sensors when closing.