Eliminate custom DeviceStatus class
Use the one provided by aiokatcp instead. The custom class used
OrderedEnum, and restrictions on Python meant it couldn't be mixed in to
the aiokatcp class. However, this functionality was only used in one
place, so it is worked around by just comparing the enum values instead
of the enums themselves.

Closes NGC-1036.