Start preparations for aiomonitor 0.6.0+
It has an extra port (for the new web UI) - see
ska-sa/katsdpservices#35. This doesn't
completely handle the switch, but makes the following preparations:

1. The master controller will allocate an extra port from Singularity
   for product controllers.
2. The product controller will use this allocated port as the default
   for the --aiomonitor-webui-port command-line option. This won't
   actually have any effect until ska-sa/katsdpservices#35 is released
   and the requirements are updated here.
3. The product controller will reserve the extra port for all the
   tasks that use katsdpservices' argparser. This means that once those
   packages are upgraded to use both the new aiomonitor and the new
   version of katsdpservices, they will expose the monitor (although
   only on localhost for now). This excludes katgpucbf which doesn't use
   katsdpservices' argparser.

The ordering needs to be done carefully, to ensure that any task running
with aiomonitor 0.6.0+ is also launched with a dynamically assigned
port, since otherwise two processes will both try to grab the default
port. That means:

- Any package can upgrade to the new katsdpservices immediately, because
  it is backwards-compatible with old aiomonitor.
- Any package upgrading aiomonitor needs to first upgrade
  katsdpservices.
- katgpucbf will need to update katsdpservices, then
  katsdpcontroller/generator.py will need to pass the new command-line
  option, then katgpucbf can upgrade aiomonitor.
- This change will need to be deployed before SDP packages can upgrade
  aiomonitor.

The new master controller will be able to safely launch old product
controllers, but not vice versa. The new product controller will be able
to run old SDP packages, but once updated to pass the command-line
option to katgpucbf, it will no longer be able to run older versions of
katgpucbf.