Sandbox: include lo as a gpucbf network
Also change the way agent_mkconfig is invoked so that lo is treated as a
single interface with connections to all three networks, instead of 3
separate interfaces. This ensures that the bandwidth can only be
allocated once instead of once per virtual network.

This was missed as part of #713 (NGC-1228).