Bump flake8 to the latest version
The version that was there is failing to run for me because
importlib-metadata no longer provides the API it expected.

The update also required a couple of issues to be fixed. The new version
doesn't seem to process old-style '# type: Foo' comments, so I updated
them to use modern type annotation syntax.