Fix check for all engines receiving good data
If all the engines had already set their status to nominal before
_wait_rx_device_status was called, it would time out and fail. That
happened because the `sensors` list was still populated, and hence the
"nothing to do" check did not take effect.

Fix it to only subscribe to sensors that aren't already nominal, so that
the `if not sensors` check is effective (and also because it's pointless
to subscribe to them).

Closes NGC-1033.