Tweak xbgpu LatestSensor data type (str -> bytes)
After bumping aiokatcp to 2.0.0 to validate sensor update values, it
seems the mirrored sensors (via SensorWatcher) are done as bytes-type
instead of the original str. This resulted in a hard failure in
aiokatcp.

This tweak should resolve this, but is done with view to change as I'm
not entirely sure how it affects live operation.

Contributes to: NGC-1062.