Reduce #sensor-status traffic during shutdown
When an engine is closed, the sensors are changed to "unreachable"
status. Change the value to the default value for the type instead of
the last reading. The value is irrelevant (because the status is not
nominal/warn/error) and the default can be much smaller, particularly
for eq gains.

Hopefully closes NGC-1440.