Major refactoring of agent_mkconfig
The primary functional difference is the addition of the --domains
argument, the separation of OS NUMA nodes from abstract "domains",
and the option to use L3 Caches to define domains. The last is not yet
tested though.

For backwards compatibility, the output attributes still refer to NUMA
nodes, and only allow for each device to be associated with a single
NUMA node.