Add ability to suppress 30s good data timeout
Add an option to develop mode to suppress the 30s good data timeout.

The context for this is that one potential reason for a task not
receiving good data within 30s of startup is a problem with the network.
We probably wouldn't want to run operations under such circumstances,
but if we're trying to debug network issues then having the product die
after 30s doesn't give enough time.

The FailReply is replaced by an error message in the logger for
traceability, but the product should not in fact just die.

Closes NGC-1571