Fix a few bugs in the schema
Firstly, I had not realised previously that there is no float type, just
number. So I had to fix that. Then the bump to 4.3 of the schema caused
a few issues in the unit tests.

Similarly, I had to rework a bit how the timeout test is done, and I
injected the shorter 1.0 second value via the new schema instead of
monkeypatching.