Add support for product schema version 4.4
This adds `taps` and `window_function` parameters to
gpucbf.antenna_channelised_voltage. Pass them through to katgpucbf. The
latter requires an update to katgpucbf.

Also change w_cutoff from positive_number to nonneg_number, so that
w_cutoff=0 is possible (which effectively disables the sinc in
generate_pfb_weights. This is done retroactively for all schema
versions, since the specification didn't say what the allowed range was.

Relates to NGC-1004.