Implement schema version 4.5
This adds the 'vlbi' attribute to 'narrowband', and passes on the
'pass_bandwidth' to katgpucbf.

Also increases the DDC_TAPS_RATIO to 24 to get better filter response.
A functional change is that the window_function will now always be
passed to katgpucbf - in practice that's harder to avoid now that the
default depends on whether 'vlbi' is specified (although it could be
done if necessary).

Closes NGC-1561.