Add ?dsim-signals request
This allows a client to issue requests to dsims without needing separate
connections to them. This also ensures that when a client receives the
reply to ?dsim-signals, the product controller is guaranteed to have
received the update of the steady-state-timestamp sensor from the dsim,
and hence a subsequent ?capture-start will latch it.

Relates to NGC-1542.