Do DNS resolution on Mesos offers
Resolve hostnames to IP addresses when offers are received, and do not
use an offer until we have the IP address. This address can then be used
in various other places that were doing their own DNS resolution.

This eliminates a number of failure modes. In particular, the endpoint
sensors are now guaranteed to be nominal.

Closes NGC-1531.