Work around a race condition in timing out getting Mesos resources
If the resources_timeout triggers in the same event loop iteration as we
get the resources we need for a launch, the exception message is not
particularly helpful (it's a TimeoutError rather than an
InsufficientResourcesError) and I think there is a chance that the
launch might at least partially happen anyway.

The race condition was always there, but I think #803 makes it much more
likely to occur because each time DNS resolution occurs we try to
launch, possibly causing a lot more event loop blockages.

This doesn't fix the problem that we're now failing to launch things
where you used to be able to; it just fixes the exception that occurs.

See NGC-1678.