Use aiokatcp sensor filtering to implement mirror_sensors
This makes mirror_sensors less total: we still get to mirror
product-level sensors like device-status and gui-urls. The former lets
us update the master-level device-status, and the latter is needed to
produce the list of GUIs on the / of the HTTP server.

This depends on https://github.com/ska-sa/aiokatcp/pull/112 (without it
I don't think anything crashes, but mirror_sensors will have no effect).

Relates to NGC-1660.